pass time.Time channel to PopWithTimer, move test
